SRAS(Django)/SRAS_Project/requirements.txt
Django>=5.2.3
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
mysqlclient>=2.1.0
Pillow>=10.0.0 
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.8.0